Added collaboration acceptance field
Added a model class, ProjectGroup, for the join table of Project and Group.
Added a collaboration acceptance field to the intermediate table of Project and Group.
Changed the model classes for Project and Group to have a collection of ProjectGroups for collaborations, instead of Projects having a list of Groups and vice versa.
Modified the ProjectsController class to accommodate the model changes.
Modified the Index page and to accommodate the model changes.
Updated ERD.